+++ src/index.ts
@@ -5,7 +5,7 @@ import AppDataSource from './config/db';
 async function main () {
 
 try {
-    await AppDataSource.initialize();
+    // await AppDataSource.initialize();
     app.listen(3000);
     console.log('Sever is listen on port', 3000);
 } catch (error) {
