+++ src/services/merchant.service.ts
@@ -1,4 +1,4 @@
-import { Merchant } from "../interfaces/webhook.interfaces"
+import { Merchant, MerchantWebhook } from "../interfaces/webhook.interfaces"
 
 export class MerchantAuthService {
     private async findMerchantByApiKey(apiKey: string): Promise <Merchant | null> {
@@ -6,29 +6,28 @@ export class MerchantAuthService {
             // This should be replaced with a mechanism to find the merchant from the db. The return statement remains
             const createDate = new Date()
             createDate.setDate(createDate.getDate() - 1) 
-                const merchantDetails = {
-                    id: crypto.randomUUID(),
-                    name: 'random-merchant',
-                    email: 'merchant-email',
-                    secret: 'merchant-secret',
-                    apiKey,
-                    isActive: false,
-                    createdAt: new Date()
-                    }
+                
             return {
-                id: merchantDetails.id,
-                apiKey: merchantDetails.apiKey,
-                secret: merchantDetails.secret,
-                name: merchantDetails.name,
-                email: merchantDetails.email,
-                isActive: merchantDetails.isActive,
-                createdAt: createDate,
-                updatedAt: createDate,
+                ...merchantDetailsWithoutApiKey,
+                apiKey,
             }
         } catch (err) {
             console.error(err)
             return null
         }
+    }
+
+    static async getMerchantById(id: string): Promise<Merchant | null>{
+        // replace with a method to find merchant from db by id
+        const date = new Date()
+        const merchant: Omit<Merchant, 'createdAt' & 'updatedAt'> = {
+            ...sampleMerchantWithoutId,
+            id,
+        }
+        if (!merchant.isActive) {
+            throw new Error('Merchant not found');
+        }
+        return merchant
     }
 
     async validateApiKey(apiKey: string): Promise<Merchant | null> {
