+++ src/services/webhook.service.ts
@@ -2,6 +2,7 @@ import axios from 'axios'
 import crypto from 'crypto'
 import { WebhookPayload, MerchantWebhook, Merchant } from '../interfaces/webhook.interfaces'
 import { validateWebhookUrl } from '../validators/webhook.validators'
+import { MerchantAuthService } from './merchant.service';
 
 
 export class WebhookService {
@@ -11,32 +12,13 @@ export class WebhookService {
         return hmac.digest('hex')
     }
 
-    private async getMerchant(id: string): Promise<Merchant | null>{
-        // replace with a method to find merchant from db by id
-        const date = new Date()
-        const merchant: Omit<Merchant, 'createdAt' & 'updatedAt'> = {
-            id,
-            apiKey: 'random-api-key',
-            name: 'random-merchant-name',
-            email: 'randomMerchant@gmail.com',
-            secret: 'merchant-webhook-secret',
-            isActive: true,
-            createdAt: new Date(),
-            updatedAt: new Date()
-        }
-        if (!merchant.isActive) {
-            throw new Error('Merchant not found');
-        }
-        return merchant
-    }
-
     private async sendWebhookNotification(
         webhookUrl: string,
         payload: WebhookPayload,
         id: string,
     ): Promise<boolean> {
         try {
-            const merchant: Merchant | null = await this.getMerchant(id)
+            const merchant: Merchant | null = await MerchantAuthService.getMerchantById(id)
             // if (!merchant) return
             const signature = await this.generateSignature(payload, merchant?.secret!);
             await axios.post(webhookUrl, payload, {
@@ -54,11 +36,24 @@ export class WebhookService {
         }
     }
 
-    async notifyPaymentUpdate (
+    static async getMerchantWebhook(merchantId: string): Promise<MerchantWebhook | null> {
+        // replace with mechanism to find merchant webhook from merchant id.
+        // in the future, another argument can be added for the transaction type
+        const merchantWebhook: MerchantWebhook = {
+            ...sampleWebhookWithoutMerchantId,
+            merchantId,
+        }
+        if (!merchantWebhook.isActive) {
+            throw new Error('Merchant web hook not found')
+        }
+        return merchantWebhook
+    }
+
+    private async notifyPaymentUpdate (
         merchantWebhook: MerchantWebhook,
         paymentDetails: Omit<WebhookPayload, 'timestamp'>
     ): Promise<boolean> {
-        const merchant = await this.getMerchant(merchantWebhook.merchantId)
+        const merchant = await MerchantAuthService.getMerchantById(merchantWebhook.merchantId)
         if (!merchantWebhook.isActive || !validateWebhookUrl(merchantWebhook.url)) {
             return false;
         }
@@ -69,5 +64,24 @@ export class WebhookService {
         }
 
         return this.sendWebhookNotification(merchantWebhook.url, webhookPayload, merchant?.secret!)
+    }
+
+    async notifyWithRetry(merchantWebhook: MerchantWebhook, webhookPayload: WebhookPayload, maxRetries = 3, delay = 3000) {
+        let attempts = 0;
+
+        while (attempts < maxRetries) {
+            try {
+                await this.notifyPaymentUpdate(merchantWebhook, webhookPayload);
+                return; // Exit if successful
+            } catch (err) {
+                attempts++;
+                console.error(`Attempt ${attempts} failed:`, err);
+
+                if (attempts < maxRetries) {
+                    await new Promise(resolve => setTimeout(resolve, delay)); // Wait before retrying
+                }
+            }
+        }
+        console.error("Failed to notify after maximum retries.");
     }
 }
